micro/book-service/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
